# factory-saas/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 厂区访客管控 SaaS API
  description: |
    Factory Visitor Control SaaS — 多租户访客手机管控系统 API。
    
    ## 认证方式
    - Cookie 认证：登录后通过 `token` HttpOnly cookie 进行身份验证
    - CSRF 保护：非安全方法（POST/PUT/DELETE）需携带 `X-CSRF-Token` 头，值与 `csrf_token` cookie 一致
    
    ## 错误格式
    所有 API 错误返回统一格式：
    ```json
    {
      "error": "ERROR_CODE",
      "message": "用户可见描述",
      "timestamp": "2024-01-01T00:00:00.000Z",
      "requestId": "uuid"
    }
    ```
  version: 2.0.0
  contact:
    name: Factory SaaS Support
servers:
  - url: http://localhost:3000
    description: 开发环境
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  schemas:
...
    Subscription:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        site_id:
          type: string
        area_name:
          type: string
        plan:
          type: string
          enum: [trial, monthly, yearly]
        status:
          type: string
          enum: [trial, active, grace, expired, suspended, cancelled]
        trial_ends_at:
          type: integer
          description: Unix timestamp (ms)
        paid_ends_at:
          type: integer
          nullable: true
        isActive:
          type: boolean
        remainingDays:
          type: number
    Site:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        name:
          type: string
        address:
          type: string
        created_at:
          type: integer
...
  parameters:
    csrfHeader:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
paths:
...
  /api/sites:
    get:
      tags: [sites]
      summary: 获取用户厂区列表
      security:
        - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  sites:
                    type: array
                    items:
                      $ref: '#/components/schemas/Site'
    post:
      tags: [sites]
      summary: 创建新厂区
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/csrfHeader'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                address:
                  type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  siteId:
                    type: string
                  subscriptionId:
                    type: string

  # ── User Subscriptions ────────────────────────────────────────
  /api/user/subscriptions:
    get:
      tags: [subscriptions]
      summary: 获取用户订阅列表
      security:
        - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscriptions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Subscription'
  /api/user/subscriptions/{subId}/bind-ip:
    post:
      tags: [subscriptions, network]
      summary: 绑定当前 IP 子网
      security:
        - cookieAuth: []
      parameters:
        - name: subId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/csrfHeader'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  wifiSubnet:
                    type: string

  # ── Network Rebind ─────────────────────────────────────────
  /api/sites/{siteId}/network-rebind-requests:
    get:
      tags: [network]
      summary: 查询重绑请求列表
      security:
        - cookieAuth: []
      parameters:
        - name: siteId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  requests:
                    type: array
                    items:
                      type: object
    post:
      tags: [network]
      summary: 提交网络重绑申请
      security:
        - cookieAuth: []
      parameters:
        - name: siteId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/csrfHeader'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                candidateSubnet:
                  type: string
                reason:
                  type: string
                evidence:
                  type: string
                candidateFingerprint:
                  type: object
                  properties:
                    publicIp:
                      type: string
                    gatewayIp:
                      type: string
                    lanSubnet:
                      type: string
                    ssid:
                      type: string
                    bssid:
                      type: string
                    gatewayMac:
                      type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  requestId:
                    type: string
                  riskScore:
                    type: integer

  # ── Orders & Payment ────────────────────────────────────────
...
  /api/user/profile:
    get:
      tags: [auth]
      summary: 获取用户详细资料
      security:
        - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  email:
                    type: string
                  name:
                    type: string
                  phone:
                    type: string
                  createdAt:
                    type: integer